add `RewardsManager`
